backend/pyproject.toml
[project]
name = "wafx-backend"
version = "1.0.0"
description = "WAFx — AI-Powered Security Operations Platform Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    "redis>=5.2.0",
    "aiokafka>=0.12.0",
    "opensearch-py>=2.8.0",
    "httpx>=0.28.0",
    "celery[redis]>=5.4.0",
    "boto3>=1.35.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "prometheus-fastapi-instrumentator>=7.0.2",
    "litellm>=1.55.0",
    "pyyaml>=6.0.2",
    "email-validator>=2.2.0",
    "slowapi>=0.1.9",
    "websockets>=14.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "testcontainers[postgres,redis,kafka]>=4.9.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
